+++ ethosu/regor/architecture/ethosu85/ethos_u85_performance.cpp
@@ -408,7 +408,7 @@ EthosU85ElementCycles EthosU85Performance::EstimateOutputCyclesPerElement(const 
         {
             activationPerfIndex = 0;
         }
-        else if ( fusedOp.type == OpType::Relu || fusedOp.type == OpType::Relu6 || fusedOp.type == OpType::ReluN1To1 )
+        else if ( fusedOp.type == OpType::Relu || fusedOp.type == OpType::Relu0To1 || fusedOp.type == OpType::Relu6 || fusedOp.type == OpType::ReluN1To1 )
         {
             activationPerfIndex = 1;
         }